Included preliminary bounds filtering
Included just the lower and upper bounds computation for bounds filtering.

+++ src/Visualization/Sequence_Visualizer.cpp
@@ -32,15 +32,15 @@ void Sequence_Visualization::plot_enabling_cons_alt_graph(std::string strFilePat
 
 	for (auto it_vtx = graph.begin(); it_vtx != graph.end(); it_vtx++)
 	{
-		uiVtx1 = it_vtx->first;
-		uiRobot1 = alt_graph.get_vertex_ownership(uiVtx1);
-		uiPos1 = alt_graph.get_vertex_position(uiVtx1);
+		uiVtx1 = (int) it_vtx->first;
+		uiRobot1 = (int)alt_graph.get_vertex_ownership(uiVtx1);
+		uiPos1 = (int) alt_graph.get_vertex_position(uiVtx1);
 
 		for (auto it_neighs = it_vtx->second.begin(); it_neighs != it_vtx->second.end(); it_neighs++)
 		{
-			uiVtx2 = it_neighs->first;
-			uiRobot2 = alt_graph.get_vertex_ownership(uiVtx2);
-			uiPos2 = alt_graph.get_vertex_position(uiVtx2);
+			uiVtx2 = (int) it_neighs->first;
+			uiRobot2 = (int) alt_graph.get_vertex_ownership(uiVtx2);
+			uiPos2 = (int) alt_graph.get_vertex_position(uiVtx2);
 
 			altGraphFile << uiPos1 << " " << uiRobot1 << " " << uiPos2 - uiPos1 << " " << uiRobot2 - uiRobot1 << "\n";
 		}
@@ -56,15 +56,15 @@ void Sequence_Visualization::plot_coll_cons_alt_graph(std::string strFilePath, c
 	
 	for (auto it_vtx = alt_graph.m_vec_vtx_alt_edge_ind_out.begin(); it_vtx != alt_graph.m_vec_vtx_alt_edge_ind_out.end(); it_vtx++)
 	{
-		uiVtx1 = it_vtx->first;
-		uiRobot1 = alt_graph.get_vertex_ownership(uiVtx1);
-		uiPos1 = alt_graph.get_vertex_position(uiVtx1);
+		uiVtx1 = (int) it_vtx->first;
+		uiRobot1 = (int) alt_graph.get_vertex_ownership(uiVtx1);
+		uiPos1 = (int) alt_graph.get_vertex_position(uiVtx1);
 
 		for (auto it_neigh = it_vtx->second.begin(); it_neigh != it_vtx->second.end(); it_neigh++)
 		{
-			uiVtx2 = it_neigh->first;
-			uiRobot2 = alt_graph.get_vertex_ownership(uiVtx2);
-			uiPos2 = alt_graph.get_vertex_position(uiVtx2);
+			uiVtx2 = (int) it_neigh->first;
+			uiRobot2 = (int) alt_graph.get_vertex_ownership(uiVtx2);
+			uiPos2 = (int) alt_graph.get_vertex_position(uiVtx2);
 
 			altGraphFile << uiPos1 << " " << uiRobot1 << " " << uiPos2 - uiPos1 << " " << uiRobot2 - uiRobot1 << "\n";
 		}
